backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
langgraph
sqlalchemy
pymysql
python-dotenv